+++ client/app/src/main/java/com/danival/game/EnergyBall.java
@@ -20,10 +20,10 @@ public class EnergyBall {
     public int type;
     public double lat;
     public double lng;
-    public long energy;
+    public int energy;
     public Circle energyUI;
 
-    public EnergyBall(MainActivity context, int _id, int _type, double _lat, double _lng, long _energy) {
+    public EnergyBall(MainActivity context, int _id, int _type, double _lat, double _lng, int _energy) {
         main = context;
         id = _id;
         type = _type;
