+++ client/app/src/main/java/com/danival/game/Food.java
@@ -21,11 +21,11 @@ public class Food {
     public int type;
     public double lat;
     public double lng;
-    public long energy;
+    public int energy;
     public Circle energyUI;
     public Marker marker;
 
-    public Food(MainActivity context, int _id, int _type, double _lat, double _lng, long _energy) {
+    public Food(MainActivity context, int _id, int _type, double _lat, double _lng, int _energy) {
         main = context;
         id = _id;
         type = _type;
